[memcached] fixed two bugs around incr command
Summary: 1. If we read how add_delta() is implemented, it has a race condition: if two incr commands come in at about the same time, they will call item_get() to get the same item say A. Now, thread 1 wants to increment the value and it may find out it needs to realloc the item, so it calls item_replace(A, B). Thread 2 may do exactly the same item_replace(A, C). Here's the problem, since A is unlinked and replaced with B in assoc table, item_replace(A, C) will have a no-op when unlinking A, then assoc table has the key, crashing insertion of C!

         2. Now, the "res + 2 > it->nbytes" test is problematic, since the slab may have given the item more space than needed. So total size of item data is NOT it->nbytes, as the code implies. This means, we were replacing items whenever item count jumps from 9 to 10, or 99 to 100, ... and most, if not all, of them are not needed.

Reviewed By: sgrimm

Test Plan: running patched code on mc010.sf2p and mc011.sf2p now.

Revert: OK




git-svn-id: http://svn.facebook.com/svnroot/projects/memcached/trunk@88203 2c7ba8d8-a2f7-0310-a573-de162e16dcc7