Don't leave stale data on failed set attempt.
This is mildly less confusing for folks with OOM errors or who store items for long periods of time, but occasionally attempt setting a very large item. While I don't believe this actually fixes a class of bugs we've seen due to refcount leaks, I make this change for consistency.