connection buffers are per-thread to avoid lock contention
Summary: - A connection buffer group is created per thread.  It gets the thread id so if asserts are compiled in, it can determine if it is being used by a different thread.
         - conn_new and dispatch_conn_new no longer needs an initial read buffer size.  it is meaningless in the context of connection buffers.
         
         Minor change:
         - udp support for stress test

Reviewed By: ps

Test Plan: ran stress test with asserts on for 1 hr with 4 clients that reconnect every 15 minutes.
           
           ran in production with asserts on for 10 minutes (and still going...)

Revert: OK


git-svn-id: http://svn.facebook.com/svnroot/projects/memcached/trunk@122288 2c7ba8d8-a2f7-0310-a573-de162e16dcc7